+++ src/Control/Comonad/Hoist/Class.hs
@@ -1,10 +1,5 @@
 {-# LANGUAGE RankNTypes #-}
-{-# LANGUAGE CPP #-}
-#if __GLASGOW_HASKELL__ >= 704
 {-# LANGUAGE Safe #-}
-#elif __GLASGOW_HASKELL__ >= 702
-{-# LANGUAGE Trustworthy #-}
-#endif
 -----------------------------------------------------------------------------
 -- |
 -- Module      :  Control.Comonad.Hoist.Class