+++ src/Control/Comonad/Traced/Class.hs
@@ -1,13 +1,8 @@
-{-# LANGUAGE CPP #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE FunctionalDependencies #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE UndecidableInstances #-}
-#if __GLASGOW_HASKELL__ >= 704
 {-# LANGUAGE Safe #-}
-#elif __GLASGOW_HASKELL__ >= 702
-{-# LANGUAGE Trustworthy #-}
-#endif
 -----------------------------------------------------------------------------
 -- |
 -- Module      :  Control.Comonad.Traced.Class
@@ -29,9 +24,6 @@ import Control.Comonad.Trans.Env
 import Control.Comonad.Trans.Store
 import qualified Control.Comonad.Trans.Traced as Traced
 import Control.Comonad.Trans.Identity
-#if __GLASGOW_HASKELL__ < 710
-import Data.Semigroup
-#endif
 
 class Comonad w => ComonadTraced m w | w -> m where
   trace :: m -> w a -> a